requirements.txt
arrow==1.4.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
circuitbreaker==2.1.3
click==8.0.4
colorama==0.4.6
cryptography==45.0.7
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
jmespath==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oci==2.164.0
packaging==25.0
pillow==12.0.0
prompt-toolkit==3.0.43
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.37.0
requests==2.32.5
responses==0.25.8
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.4
terminaltables==3.1.10
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.0
wcwidth==0.2.14
